frontend/requirements.txt
streamlit==0.86.0
python-dotenv==0.19.0
matplotlib==3.4.2
plotly==5.1.0
